detection-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
requests==2.31.0
python-whois==0.8.0
beautifulsoup4==4.12.2
ssdeep==3.4
pydantic==2.5.0
python-dotenv==1.0.0